agent/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
PyMySQL==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
